+++ example/app/app/factories/DefaultEventsFactory.swift
@@ -1,10 +1,3 @@
-//
-//  DefaultEventsFactory.swift
-//  app
-//
-//  Created by Sergey Korney
-//
-
 import AffiseAttributionLib