+++ example/app/app/factories/EventsFactory.swift
@@ -1,10 +1,3 @@
-//
-//  EventsFactory.swift
-//  app
-//
-//  Created by Sergey Korney
-//
-
 import AffiseAttributionLib
 
 protocol EventsFactory {